# src/main/resources/static/docs/openapi3.yaml
openapi: 3.0.1
info:
  title: API 문서
  description: RestDocs With Swagger
  version: 0.0.1
servers:
- url: http://devterview.com:8080
- url: http://localhost:8080
- url: http://dev-api.devterview.com:8080
- url: http://dev-api.devterview.com
...
paths:
...
  /file/cover-letter:
    get:
      tags:
      - File API
      summary: 파일 API
      description: 파일 API
      operationId: 유저가 올린 자소서 목록 조회
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/file-cover-letter1504893150'
              examples:
                유저가 올린 자소서 목록 조회:
                  value: |-
                    {
                      "code" : 200,
                      "message" : "SUCCESS",
                      "data" : {
                        "coverLetters" : [ {
                          "fileId" : 1,
                          "type" : "COVER_LETTER",
                          "fileName" : "testFile.txt",
                          "s3FileUrl" : "https://s3.aws.com/testFile.txt"
                        }, {
                          "fileId" : 2,
                          "type" : "COVER_LETTER",
                          "fileName" : "testFile2.txt",
                          "s3FileUrl" : "https://s3.aws.com/testFile2.txt"
                        }, {
                          "fileId" : 3,
                          "type" : "COVER_LETTER",
                          "fileName" : "testFile3.txt",
                          "s3FileUrl" : "https://s3.aws.com/testFile3.txt"
                        } ]
                      }
                    }
    post:
      tags:
      - File API
      summary: 마이페이지에서 자소서 저장
      description: 마이페이지에서 자소서 저장
      operationId: 마이페이지에서 자소서 저장
      requestBody:
        content:
          application/json;charset=UTF-8:
            schema:
              $ref: '#/components/schemas/file-cover-letter-412892686'
            examples:
              마이페이지에서 자소서 저장:
                value: |-
                  {
                    "type" : "COVER_LETTER",
                    "filePath" : "filePath"
                  }
      responses:
        "201":
          description: "201"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/file-cover-letter-1298800292'
              examples:
                마이페이지에서 자소서 저장:
                  value: |-
                    {
                      "code" : 200,
                      "message" : "SUCCESS",
                      "data" : {
                        "fileId" : 2,
                        "type" : "COVER_LETTER",
                        "fileName" : "fileName",
                        "s3FileUrl" : "s3FileUrl"
                      }
                    }
...
  /file/cover-letter/{fileName}/upload-url:
    get:
      tags:
      - File API
      summary: 파일 API
      description: 파일 API
      operationId: 마이페이지에서 PreSigned URL 생성
      parameters:
      - name: fileName
        in: path
        description: 업로드할 파일명
        required: true
        schema:
          type: string
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/file-cover-letter-fileName-upload-url803243778'
              examples:
                마이페이지에서 PreSigned URL 생성:
                  value: |-
                    {
                      "code" : 200,
                      "message" : "SUCCESS",
                      "data" : {
                        "preSignedUrl" : "https://s3.aws.com/testFile.txt",
                        "objectKey" : "testFile.txt"
                      }
                    }
  /file/cover-letter/{interviewId}/download-url:
    get:
      tags:
      - File API
      summary: 파일 API
      description: 파일 API
      operationId: 면접 자소서 다운로드 url 받기 - 성공
      parameters:
      - name: interviewId
        in: path
        description: ""
        required: true
        schema:
          type: string
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/file-cover-letter-interviewId-download-url-1410657502'
              examples:
                면접 자소서 다운로드 url 받기 - 성공:
                  value: |-
                    {
                      "code" : 200,
                      "message" : "SUCCESS",
                      "data" : {
                        "preSignedUrl" : "https://s3.aws.com/cover-letter/1/cover-letter-test.txt",
                        "objectKey" : "/files/cover-letter/2/1/cover-letter-test.txt"
                      }
                    }
  /file/profile-image/{fileName}/upload-url:
    get:
      tags:
      - File API
      summary: 파일 API
      description: 파일 API
      operationId: 프로필 이미지 PreSignedUrl 받기 - 성공
      parameters:
      - name: fileName
        in: path
        description: 업로드할 파일명
        required: true
        schema:
          type: string
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/file-cover-letter-interviewId-download-url-1410657502'
              examples:
                프로필 이미지 PreSignedUrl 받기 - 성공:
                  value: |-
                    {
                      "code" : 200,
                      "message" : "SUCCESS",
                      "data" : {
                        "preSignedUrl" : "https://s3.aws.com/profile-image/1/profile-image.jpg",
                        "objectKey" : "profile-img.jpg"
                      }
                    }
  /file/audio/{interviewId}/{questionId}/download-url:
    get:
      tags:
      - File API
      summary: 오디오 파일 다운로드를 위한 PreSigned URL 생성
      description: 오디오 파일 다운로드를 위한 PreSigned URL 생성
      operationId: 오디오 파일 다운로드 PreSigned URL 생성 - 성공
      parameters:
      - name: interviewId
        in: path
        description: 면접 ID
        required: true
        schema:
          type: string
      - name: questionId
        in: path
        description: 질문 ID
        required: true
        schema:
          type: string
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/file-audio-interviewId-questionId-download-url-37488830'
              examples:
                오디오 파일 다운로드 PreSigned URL 생성 - 성공:
                  value: |-
                    {
                      "code" : 200,
                      "message" : "SUCCESS",
                      "data" : {
                        "preSignedUrl" : "https://s3.aws.com/audio/1/interview/2/question/3/audio-test.mp3",
                        "objectKey" : "audio/1/interview/2/question/3/audio-test.mp3"
                      }
                    }
  /file/audio/{interviewId}/{questionId}/upload-url:
    get:
      tags:
      - File API
      summary: 오디오 파일 업로드를 위한 PreSigned URL 생성
      description: 오디오 파일 업로드를 위한 PreSigned URL 생성
      operationId: 오디오 파일 업로드 PreSigned URL 생성
      parameters:
      - name: interviewId
        in: path
        description: 면접 ID
        required: true
        schema:
          type: string
      - name: questionId
        in: path
        description: 질문 ID
        required: true
        schema:
          type: string
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/file-audio-interviewId-questionId-download-url-37488830'
              examples:
                오디오 파일 업로드 PreSigned URL 생성:
                  value: |-
                    {
                      "code" : 200,
                      "message" : "SUCCESS",
                      "data" : {
                        "preSignedUrl" : "https://s3.aws.com/audio/1/interview/2/question/3/audio-test.mp3",
                        "objectKey" : "audio/1/interview/2/question/3/audio-test.mp3"
                      }
                    }
  /file/cover-letter/user/{coverLetterId}/download-url:
    get:
      tags:
      - File API
      summary: 파일 API
      description: 파일 API
      operationId: 유저 자소서 다운로드 url 받기 - 성공
      parameters:
      - name: coverLetterId
        in: path
        description: ""
        required: true
        schema:
          type: string
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/file-cover-letter-interviewId-download-url-1410657502'
              examples:
                유저 자소서 다운로드 url 받기 - 성공:
                  value: |-
                    {
                      "code" : 200,
                      "message" : "SUCCESS",
                      "data" : {
                        "preSignedUrl" : "https://s3.aws.com/9e715748-52d8-49b2-96dd-a436e9b15903.docx",
                        "objectKey" : "9e715748-52d8-49b2-96dd-a436e9b15903.docx"
                      }
                    }
  /file/cover-letter/{interviewId}/{fileName}/upload-url:
    get:
      tags:
      - File API
      summary: 파일 API
      description: 파일 API
      operationId: 면접 정보 입력 시 PreSigned URL 생성
      parameters:
      - name: interviewId
        in: path
        description: 면접 ID
        required: true
        schema:
          type: string
      - name: fileName
        in: path
        description: 업로드할 파일명
        required: true
        schema:
          type: string
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/file-cover-letter-interviewId-download-url-1410657502'
              examples:
                면접 정보 입력 시 PreSigned URL 생성:
                  value: |-
                    {
                      "code" : 200,
                      "message" : "SUCCESS",
                      "data" : {
                        "preSignedUrl" : "https://s3.aws.com/testFile.txt",
                        "objectKey" : "testFile.txt"
                      }
                    }
...
components:
  schemas:
...
    file-audio-interviewId-questionId-download-url-37488830:
      type: object
      properties:
        code:
          type: number
          description: 응답 코드
        data:
          type: object
          properties:
            objectKey:
              type: string
              description: aws s3 object key
            preSignedUrl:
              type: string
              description: AWS S3 PreSigned URL
        message:
          type: string
          description: 응답 메시지
...
    file-cover-letter-412892686:
      type: object
      properties:
        filePath:
          type: string
          description: s3 파일 경로
        type:
          type: string
          description: "자소서 타입: COVER_LETTER"
...
    file-cover-letter-interviewId-download-url-1410657502:
      type: object
      properties:
        code:
          type: number
          description: 응답 코드
        data:
          type: object
          properties:
            objectKey:
              type: string
              description: aws s3 object key
            preSignedUrl:
              type: string
              description: aws s3 preSigned Url
        message:
          type: string
          description: 응답 메시지
...
    file-cover-letter-1298800292:
      type: object
      properties:
        code:
          type: number
          description: 응답 코드
        data:
          type: object
          properties:
            fileName:
              type: string
              description: 파일 이름
            s3FileUrl:
              type: string
              description: 파일 경로
            type:
              type: string
              description: 파일 타입
            fileId:
              type: number
              description: 파일 식별자
        message:
          type: string
          description: 응답 메시지
...
    file-cover-letter-fileName-upload-url803243778:
      type: object
      properties:
        code:
          type: number
          description: 응답 코드
        data:
          type: object
          properties:
            objectKey:
              type: string
              description: aws s3 object key
            preSignedUrl:
              type: string
              description: aws s3 PreSigned Url
        message:
          type: string
          description: 응답 메시지
...
    file-cover-letter1504893150:
      type: object
      properties:
        code:
          type: number
          description: 응답 코드
        data:
          type: object
          properties:
            coverLetters:
              type: array
              items:
                type: object
                properties:
                  fileName:
                    type: string
                    description: 자소서 파일 이름
                  s3FileUrl:
                    type: string
                    description: 자소서 업로드 주소
                  type:
                    type: string
                    description: "파일 유형: COVER_LETTER"
                  fileId:
                    type: number
                    description: 자소서 식별자
        message:
          type: string
          description: 응답 메시지
...
  securitySchemes:
    APIKey:
      type: apiKey
      name: access_token
      in: cookie
security:
  - APIKey: []